requirements.txt
alembic==1.16.4
blinker==1.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
